server app routing and refactorization
